Change for test case structure
